% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{atlas_ui}
\alias{atlas_ui}
\title{Build the Shiny UI for the Atlas Explorer}
\usage{
atlas_ui(metadata_choices)
}
\arguments{
\item{metadata_choices}{Character vector of metadata column names.}
}
\value{
A Shiny UI definition.
}
\description{
Build the Shiny UI for the Atlas Explorer
}
\keyword{internal}
