% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_metadata_choices}
\alias{get_metadata_choices}
\title{Extract suitable metadata columns from a Seurat object}
\usage{
get_metadata_choices(obj)
}
\arguments{
\item{obj}{A Seurat object.}
}
\value{
Character vector of column names.
}
\description{
Extract suitable metadata columns from a Seurat object
}
\keyword{internal}
