% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_enrich_plot}
\alias{make_enrich_plot}
\title{Build an EnrichR bar plot}
\usage{
make_enrich_plot(df, color_high, title_txt, n_terms, pt)
}
\arguments{
\item{df}{Tidy EnrichR data frame.}

\item{color_high}{High colour for fill gradient.}

\item{title_txt}{Plot title.}

\item{n_terms}{Number of terms.}

\item{pt}{Base font size.}
}
\value{
A ggplot object.
}
\description{
Build an EnrichR bar plot
}
\keyword{internal}
