% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_size_controls}
\alias{plot_size_controls}
\title{Sidebar plot-size control block}
\usage{
plot_size_controls(
  id_prefix,
  default_w = 800,
  default_h = 560,
  default_pt = 12
)
}
\arguments{
\item{id_prefix}{Character prefix for input IDs.}

\item{default_w, default_h}{Default display width/height in pixels.}

\item{default_pt}{Default font size in points.}
}
\value{
A tagList of Shiny UI elements.
}
\description{
Sidebar plot-size control block
}
\keyword{internal}
