% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{run_enrichr_tidy}
\alias{run_enrichr_tidy}
\title{Run EnrichR and return a tidy data frame}
\usage{
run_enrichr_tidy(genes, db)
}
\arguments{
\item{genes}{Character vector of gene symbols.}

\item{db}{EnrichR database name.}
}
\value{
A data frame or NULL.
}
\description{
Run EnrichR and return a tidy data frame
}
\keyword{internal}
